+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Changelog
+
+## [Unreleased]
+### Added
+- `--force-enac` flag to force ENAC to 1 after calibration
+
 ## [0.1.2] - 2026-06-02
 ### Added
 - CHANGELOG.md file